// src/services/generated/model/ticketCreateAndProcessPayload.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Lexxadata Customer Scraper API
 * A structured API to search and scrape customer data from the NMS portal.
 * OpenAPI spec version: 1.0.0
 */

export interface TicketCreateAndProcessPayload {
  query: string;
  description: string;
  priority?: string;
  jenis?: string;
  noc_username: string;
  noc_password: string;
}