// src/services/generated/model/ticketProcessPayload.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Lexxadata Customer Scraper API
 * A structured API to search and scrape customer data from the NMS portal.
 * OpenAPI spec version: 1.0.0
 */

export interface TicketProcessPayload {
  query: string;
  noc_username: string;
  noc_password: string;
}