// src/services/generated/model/ticketForwardPayload.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Lexxadata Customer Scraper API
 * A structured API to search and scrape customer data from the NMS portal.
 * OpenAPI spec version: 1.0.0
 */

export interface TicketForwardPayload {
  query: string;
  service_impact: string;
  root_cause: string;
  network_impact: string;
  recomended_action: string;
  onu_index: string;
  sn_modem: string;
  priority?: string;
  person_in_charge?: string;
  noc_username: string;
  noc_password: string;
}