// src/services/generated/model/ticketClosePayload.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Lexxadata Customer Scraper API
 * A structured API to search and scrape customer data from the NMS portal.
 * OpenAPI spec version: 1.0.0
 */

export interface TicketClosePayload {
  query: string;
  close_reason: string;
  onu_sn: string;
  noc_username: string;
  noc_password: string;
}